Alteração de Produto para novas regras de negocio

+++ back/src/main/java/com/gestao/back/service/ProdutoServiceImpl.java
@@ -71,6 +71,7 @@ public class ProdutoServiceImpl {
         produto.setCategoria(dto.getCategoria());
         produto.setPrecoUnitario(dto.getPrecoUnitario());
         produto.setQuantidadeEstoque(dto.getQuantidadeEstoque());
+        produto.setAtivo(true);
 
         Produto produtoSalvo = produtoRepository.save(produto);
 
@@ -90,6 +91,7 @@ public class ProdutoServiceImpl {
         produto.setNome(dto.getNome());
         produto.setCategoria(dto.getCategoria());
         produto.setPrecoUnitario(dto.getPrecoUnitario());
+        produto.setAtivo(dto.isAtivo());
 
         Produto produtoAtualizado = produtoRepository.save(produto);
         return new ProdutoResponseDTO(produtoAtualizado);
@@ -97,8 +99,14 @@ public class ProdutoServiceImpl {
 
     @Transactional
     public void deletarProduto(Long id) {
+        if(!produtoRepository.getReferenceById(id).isAtivo()) {
+            throw new IllegalArgumentException("Produto precisa estar desativado para ser deletado");
+        }
         if (!produtoRepository.existsById(id)) {
             throw new EntityNotFoundException("Produto não encontrado");
+        }
+        if(produtoRepository.getReferenceById(id).getQuantidadeEstoque() != 0){
+            throw new IllegalArgumentException("Produto não pode ser deletado pois esta em estoque");
         }
         produtoRepository.deleteById(id);
     }
